print an error and the title of the page when clifox connects to an instance of firefox whose browsing context (gBrowser) is unavailable
Firefox will insert its own dialogs from time to time at the start of a firefox session.
Now, the title of this popup dialog will be presented to the user, and clifox will exit.
The user can send this title to the developers to get a prefs.js patch to remove said dialog, or an application patch to close this dialog programatically.