add fix for "missing" end node markers, add autoswitching tab functionality
When a webpage has nodes at its end that are considered ignorable, mainly script and style elements and the like, the renderer skips them
Previously, however, the renderer did not check for the end of the page, so would exceed the pages array bounds.
Now, when a user opens a page in a new window, the page is forced into a new tab, that should focus itself.