change pageLoad notification to use onStateChange (todo:insure currentBrowser==event.browser)
Because Firefox does not fire onload events when moving forward and backward through all pages, we have switched to using the onStateChange event.
This event is only sent to python when the page is marked as finished loading, the vent is marked as is_window, and the window referred to is a top-level window.