add gotoBookmark function for selectign and navigating to a bookmark
By default, the gotoBookmark function is set to the F12 key, in conf/clifox.conf.
When the gotoBookmark key is pressed, a treeview will open.
This treeview presents the user with a list of bookmark folders.
Use the down arrow and up arrow to move to the next or previous entry.
Use the right arrow to expand a folder.
(If a folder has contents, they will be displayed underneath the parent that opened them.)
When in a list of bookmarks, or "children", pressing the left arrow once will move you to the parent of the selected item.
Pressing the arrow a second time, or for the first time on an open branch, collapses that branch and all it's children.
Pressing enter on an item returns that item to the caller, if it is selectable. (no action will occur if the selected item is not an appropriate type (e.g. a bookmark folder, rather than a bookmark itself).)
Pressing q exits the treeview.