add improved accessible parser class, make tab switching asynchronis
For the sake of uniformity and to work with event-driven API's in firefox, window switching now sets a variable in firefox; firefox then sends an event to clifox.
Clifox then acts on that event.
Most future access to webpages will be done asyncronously (e.g. nsIAccessible page change events, ajax, etc).
This change permits interoperability with the above mentioned code.