fix showScreen to be more intuitive
When showScreen is called, it checks first to see if the caller provided absoluteY, the vertical index of a webpage.
If so, it uses absoluteY to determine the other values, screenNumber, etc.
If absoluteY is not provided, showScreen uses the provided screenNumber and vertical offset inside that screen,
defaulting to the current cursor position of said values are not provided.