change js objects to less array-based locator
Instead of searching an array to check if an object is already in the list of possible remote objects, now attempt to check the object itself, before resorting to an array search.
This should speed up cross-language access somewhat, especially when there are more than twenty-thousand js objects (possible when we have multiple pages open or nested frames).